# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ArtifactFlow API",
    "description": "Multi-agent system API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "description": "\u53d1\u9001\u65b0\u6d88\u606f\uff08multipart/form-data\uff09\n\n`payload` \u4e3a ChatRequest \u7684 JSON \u5b57\u7b26\u4e32\uff1b`files` \u4e3a\u53ef\u9009\u9644\u4ef6\u3002\u9644\u4ef6\u5728\u8d77 turn\n\u524d\u540c\u6b65\u8f6c\u6210 artifact\uff08source=user_upload\uff09\u843d\u5e93\uff0c\u5e76\u628a (id, filename) \u900f\u4f20\u8fdb\nUSER_INPUT \u4e8b\u4ef6\u6b63\u6587\uff0c\u8ba9 agent \u77e5\u9053\u54ea\u4e9b\u662f\u672c\u8f6e\u65b0\u4f20\u7684\u3002\u8fd4\u56de stream_url \u4f9b\u524d\u7aef\u8ba2\u9605\u3002",
        "operationId": "send_message_api_v1_chat_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_message_api_v1_chat_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Conversations",
        "description": "\u5217\u51fa\u5bf9\u8bdd\u5217\u8868",
        "operationId": "list_conversations_api_v1_chat_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/{conv_id}/active-stream": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Active Stream",
        "description": "\u67e5\u8be2\u4f1a\u8bdd\u662f\u5426\u6709\u6d3b\u8dc3\u7684\u6267\u884c\u6d41\uff0c\u7528\u4e8e\u65ad\u7ebf\u91cd\u8fde\u3002\u65e0\u6d3b\u8dc3\u6d41\u662f\u6b63\u5e38\u7a7a\u72b6\u6001\u3002",
        "operationId": "get_active_stream_api_v1_chat__conv_id__active_stream_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveStreamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/{conv_id}/inject": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Inject Message",
        "description": "\u5411\u6d3b\u8dc3\u6267\u884c\u6ce8\u5165\u6d88\u606f\n\n\u4ec5\u5f53 conversation \u6709\u6b63\u5728\u8fd0\u884c\u7684\u6267\u884c\u65f6\u53ef\u7528\u3002\n\u6ce8\u5165\u7684\u6d88\u606f\u901a\u8fc7 queued_message \u4e8b\u4ef6\u8fdb\u5165 lead agent \u7684 context\u3002\n\u524d\u7aef\u4e0d\u5e94\u91cd\u5efa SSE \u8fde\u63a5 \u2014 \u4e8b\u4ef6\u4ecd\u901a\u8fc7\u539f\u6709 stream \u63a8\u9001\u3002\n\n\u6ce8\u5165\u5185\u5bb9\u901a\u8fc7 queued_message \u4e8b\u4ef6\u6301\u4e45\u5316\u5230 MessageEvent \u8868\uff0c\n\u53ef\u901a\u8fc7 GET /chat/{conv_id}/messages/{msg_id}/events \u67e5\u8be2\u3002\n\u4e0d\u521b\u5efa\u72ec\u7acb Message \u8bb0\u5f55\uff08\u6ce8\u5165\u662f\u540c\u4e00\u8f6e\u6267\u884c\u7684\u8865\u5145\u8f93\u5165\uff0c\u975e\u72ec\u7acb\u5bf9\u8bdd\u8f6e\u6b21\uff09\u3002",
        "operationId": "inject_message_api_v1_chat__conv_id__inject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/{conv_id}/cancel": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Cancel Execution",
        "description": "\u53d6\u6d88\u6d3b\u8dc3\u6267\u884c\n\n\u8bf7\u6c42\u53d6\u6d88 conversation \u5f53\u524d\u6b63\u5728\u8fd0\u884c\u7684\u6267\u884c\u3002\u5f15\u64ce\u4f1a\u5728\u4e0b\u4e00\u4e2a\u68c0\u67e5\u70b9\u4f18\u96c5\u9000\u51fa\u3002",
        "operationId": "cancel_execution_api_v1_chat__conv_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/storage": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Storage Usage",
        "description": "\u5f53\u524d\u7528\u6237\u7684\u9644\u4ef6\u5b58\u50a8\u7528\u91cf + \u914d\u989d\uff08\u5582\u524d\u7aef\u8fdb\u5ea6\u6761\uff09\u3002\n\n\u4e0e\u4e0a\u4f20\u6321\u677f\u540c\u53e3\u5f84\uff08compute-on-read\uff0c\u5355\u4e00\u6570\u636e\u6e90\uff09\u3002\u58f0\u660e\u5728 `/{conv_id}` \u4e4b\u524d\uff0c\n\u5426\u5219 `storage` \u4f1a\u88ab\u5f53\u4f5c conv_id \u547d\u4e2d\u8be6\u60c5\u8def\u7531\u3002",
        "operationId": "get_storage_usage_api_v1_chat_storage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageUsageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_send_message_api_v1_chat_post": {
        "properties": {
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "payload",
          "files"
        ],
        "title": "Body_send_message_api_v1_chat_post"
      },
...
      "ChatResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "description": "Conversation ID"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id",
            "description": "New message ID"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url",
            "description": "SSE endpoint URL for streaming"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "message_id",
          "stream_url"
        ],
        "title": "ChatResponse",
        "description": "POST /api/v1/chat response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "required": [
          "detail"
        ]
      },
      "InjectRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 20000,
            "title": "Content",
            "description": "Message content to inject into the active execution"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "InjectRequest",
        "description": "POST /api/v1/chat/{conv_id}/inject request body"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}